added color utils

+++ src/main/java/org/eomasters/utils/Colors.java
@@ -1,3 +1,20 @@
+/*-
+ * ========================LICENSE_START=================================
+ * EOM Commons - Library of common utilities for Java
+ * -> https://www.eomasters.org/
+ * ======================================================================
+ * Copyright (C) 2023 - 2024 Marco Peters
+ * ======================================================================
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ * =========================LICENSE_END==================================
+ */
+
 package org.eomasters.utils;
 
 import java.awt.Color;
@@ -11,7 +28,7 @@ public class Colors {
    * @param alpha the alpha value
    * @return the created color
    */
-  public static Color create(String rgb, float alpha) {
+  public static Color create(String rgb, double alpha) {
     if (rgb.length() > 7) {
       rgb = rgb.substring(0, 7);
     }
